requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.14
django-phonenumber-field==8.0.0
django-recaptcha==4.0.0
idna==3.7
phonenumberslite==8.13.40
pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
urllib3==2.2.2
